Add support for the removeAboveValue and removeBelowValue functions
The removeAboveValue(seriesList, n) function removes data
above the given threshold from a series or list of series.
Values above this threshold are assigned a value of 'null'.

The removeBelowValue(seriesList, n) function removes data
below the given threshold from a series or list of series.
Values below this threshold are assigned a value of 'null'.