Add support for the diffSeries function
The diffSeries function subtracts metrics, lists of metrics and
constant values (parameters 2 through n) from parameter 1 - which
in turn is fleixible and could be a metric, list of metrics or a
constant value. Order of parameters 2 through n is not important.
Where as parameter 1 provides the initial value or list of values
to subtract from

Examples of usage:

    diffSeries(5, foo.bar);
    diffSeries(foo.*, 2, meep.bar);
    diffSeries(*.*);