Metrics are now defined through a concept of 'groups'
A group defines 1 or more metrics (a metric can be a wildcard) who all share
common rendering behaviour (i.e. are lines, or areas) and are associated
with the same axis.

The order of groups within overall definition determines 'depth', with the
lower numbered groups being rendered 'first'.

Currently hard-coded to only support 2 groups, but would be relatively
easy to change in the future to support unlimited groupings.

At this commit, there is still only one axis (left) and only the line
interpolation changes are used.