Add support for the movingAverage function
The moving average function charts the moving average of a metric
or metrics liset over a fixed number of   past points, or a time interval.

It does this by taking one metric or a wildcard seriesList followed by a
number of datapoints or a quoted string with a length of time like '1hour' or '5min'.

Calculating the average of the preceeding datapoints for each point on the graph.

All previous datapoints are set to 'None' at the beginning of the graph.